+++ contracts/mocks/StartTokenIdHelper.sol
@@ -10,9 +10,9 @@ pragma solidity ^0.8.4;
  * to be returned by the overriden `_startTokenId()` function of ERC721A in the ERC721AStartTokenId mocks.
  */
 contract StartTokenIdHelper {
-    uint256 public startTokenId;
+    uint128 public startTokenId;
 
-    constructor(uint256 startTokenId_) {
+    constructor(uint128 startTokenId_) {
         startTokenId = startTokenId_;
     }
 }
